refactor(models): strip UI helpers from bus models into extensions
