+++ lib/models/bus_violation_records_data.dart
@@ -1,9 +1,6 @@
 import 'dart:convert';
 
-import 'package:flutter/cupertino.dart';
 import 'package:json_annotation/json_annotation.dart';
-import 'package:nkust_ap/l10n/nkust_localizations.dart';
-import 'package:nkust_ap/utils/utils.dart';
 
 part 'bus_violation_records_data.g.dart';
 
@@ -75,14 +72,6 @@ class Reservation {
   });
 
   String get amountendText => (amountend == 0) ? '' : '\$$amountend';
-
-  String? startStationText(BuildContext context) {
-    return Utils.parserCampus(context.t, startStation);
-  }
-
-  String? endStationText(NkustLocalizations local) {
-    return Utils.parserCampus(local, endStation);
-  }
 
   factory Reservation.fromJson(Map<String, dynamic> json) =>
       _$ReservationFromJson(json);
