+++ lib/pages/bus/bus_reservations_page.dart
@@ -2,6 +2,7 @@ import 'package:ap_common/ap_common.dart';
 import 'package:flutter/material.dart';
 import 'package:nkust_ap/api/exceptions/api_exception.dart';
 import 'package:nkust_ap/api/exceptions/api_exception_l10n.dart';
+import 'package:nkust_ap/extensions/bus_reservations_data_ui_extension.dart';
 import 'package:nkust_ap/models/models.dart';
 import 'package:nkust_ap/utils/global.dart';
